Casino::Listar
Listar para ficheiro a funcionar, dar volta ao cosido para melhorar estrutura

+++ TrabalhoFinal/include/MBlackJack.h
@@ -8,7 +8,7 @@ class MBlackJack : public Maquina
 {
 
     public:
-        MBlackJack(int _nM, const std::string& _nome, const std::string& _x, const std::string& _y, int _premio, int _prob, const std::string& _tipo);
+        MBlackJack(int _nM, const std::string& _nome, const std::string& _x, const std::string& _y, int _premio, float _prob, const std::string& _tipo);
         virtual ~MBlackJack();
 
     protected:
